cheat_detect_server/requirements.txt
absl-py==2.0.0
annotated-types==0.6.0
anyio==4.2.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.0
cycler==0.12.1
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.108.0
filelock==3.13.1
filterpy==1.4.5
flatbuffers==23.5.26
fonttools==4.44.3
fsspec==2023.10.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
grpcio==1.59.3
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
humanfriendly==10.0
idna==3.6
imageio==2.33.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
lap==0.4.0
lazy_loader==0.3
loguru==0.7.2
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
motmetrics==1.4.0
MouseInfo==0.1.3
mpmath==1.3.0
netifaces==0.11.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnx==1.15.0
onnx-simplifier==0.4.35
onnxruntime==1.16.2
opencv-python==4.8.1.78
orjson==3.9.10
packaging==23.2
pandas==2.1.3
pillow==10.2.0
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pydantic==2.5.3
pydantic-extra-types==2.3.0
pydantic-settings==2.1.0
pydantic_core==2.14.6
PyGetWindow==0.0.9
Pygments==2.17.1
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
python3-xlib==0.15
pytweening==1.0.7
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
scikit-image==0.22.0
scipy==1.11.4
sniffio==1.3.0
starlette==0.32.0.post1
sympy==1.12
tabulate==0.9.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
tifffile==2023.9.26
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.9.0
urllib3==2.1.0
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
xmltodict==0.13.0
